requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.2.1
django-allauth==65.7.0
django-cors-headers==4.7.0
django-csp==4.0
djangorestframework==3.12.4
djongo==1.3.7
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
oauthlib==3.2.2
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pymongo==3.11.4
python-telegram-bot==22.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
whitenoise==6.9.0
